webSys/requirements.txt
alembic==0.9.1
appdirs==1.4.3
asn1crypto==0.22.0
Babel==2.4.0
bcolz==1.1.2
better-exceptions==0.1.8
blinker==1.4
cffi==1.10.0
click==6.7
colorama==0.3.9
cryptography==1.8.1
cycler==0.10.0
decorator==4.0.11
dill==0.2.5
ez-setup==0.9
feedparser==5.2.1
Flask==0.12.1
Flask-Babel==0.11.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
flask-paginate==0.4.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
idna==2.5
ipython==6.0.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jsonpickle==0.9.4
line-profiler==2.0
Logbook==1.0.0
lxml==3.7.3
Mako==1.0.6
MarkupSafe==1.0
matplotlib==2.0.2
misaka==2.1.0
mysqlclient==1.3.10
numpy==1.12.1
olefile==0.44
packaging==16.8
pandas==0.19.2
pickleshare==0.7.4
Pillow==4.1.0
prompt-toolkit==1.0.14
py==1.4.33
pycparser==2.17
Pygments==2.2.0
PyMySQL==0.7.11
pyparsing==2.2.0
pyreadline==2.1
pytest==3.1.0
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2017.2
PyYAML==3.12
requests==2.13.0
rqalpha==2.2.7
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
SQLAlchemy==1.1.9
tabulate==0.7.7
tornado==4.5.1
traitlets==4.3.2
tushare==0.7.4
wcwidth==0.1.7
Werkzeug==0.12.1
win-unicode-console==0.5
XlsxWriter==0.9.6
yahoo-finance==1.4.0
